Feat:모임 이미지  수정
