--- src/components/common/FileInputs/FileInputs.module.scss
@@ -1,24 +1,0 @@
-.imageUploadBox {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-@mixin imagePreview {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  height: 100%;
-}
-
-.imagePreview {
-  @include imagePreview;
-}
-
-.defaultImage {
-  @include imagePreview;
-  position: relative;
-  border-radius: 10px;
-  background: var(--lukyviky-gray-Light, #f5f6f7);
-}
